requirements.txt
pandas==1.1.5
numpy==1.19.5
streamlit==0.71.0
pycaret==2.2.3
Pillow==8.0.1
